Fix ratio limit in documentation
The value of ratio isn't limited to 0-1, which can be checked in the two first links below. It probably was a mistake caused by copy-pasting the documentation of "progress", which is limited to 0-1. The commit where this extra info was added is the one in the third link.

- https://github.com/webtorrent/webtorrent/blob/b53d224cdea46173ea97dc9eea7eaf5a6e7a55b9/lib/torrent.js#L177-L179
- https://github.com/webtorrent/webtorrent/blob/b53d224cdea46173ea97dc9eea7eaf5a6e7a55b9/index.js#L195-L205
- https://github.com/webtorrent/webtorrent/commit/6f2e85cd4649b3f1ddd67be9f4832c0138849b3e

As this value is the result of something like `uploaded / max(1, downloaded)` it never get's to `Infinity`, so it's limited by 0 and `uploaded`. Without the ` || 1` part it would be limited between 0 and infinity.